+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- ability to send plain strings with `NostrTemplate`
+
 ## [0.0.1] - 2024-04-25
 
 ### Added